pyproject.toml
[tool.poetry]
name = "python-poetry-gemeni"
version = "0.1.0"
description = ""
authors = ["masfana016 <masfanasrullahansari123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
langchain = "0.3.4"
langchain-google-community = "^2.0.1"
langchain-google-genai = "2.0.1"
langchain-tools = "^0.1.34"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
